03_hello_agent/async_agent/hello_agent-app/pyproject.toml
[project]
name = "hello-agent-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.86.0",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]
